edge/requirements.txt
paho-mqtt
influxdb-client
pandas
prophet
python-dotenv